Make stats modular
