+++ mite_selenium/__init__.py
@@ -1,9 +1,28 @@
 import logging
 from contextlib import asynccontextmanager
-from selenium.webdriver import Remote
+
+from mite.stats import Counter, Histogram, Stats, label_extractor, matcher_by_type
 from mite.utils import spec_import
+from selenium.webdriver import Remote
 
 logger = logging.getLogger(__name__)
+
+
+_MITE_STATS = (
+    Histogram(
+        'mite_http_selenium_response_time_seconds',
+        matcher_by_type('http_selenium_metrics'),
+        label_extractor=label_extractor(['transaction']),
+        value_extractor=lambda x: x['total_time'],
+        bins=[0.0001, 0.001, 0.01, 0.05, 0.1, 0.2, 0.4, 0.8, 1, 2, 4, 8, 16, 32, 64],
+    ),
+    Counter(
+        'mite_http_selenium_response_total',
+        matcher_by_type('http_selenium_metrics'),
+        label_extractor('test journey transaction'.split()),
+    ),
+)
+Stats.register(_MITE_STATS)
 
 
 class _SeleniumWrapper:
